move functions to mapfunctions
