+++ src/engine/Map.hxx
@@ -37,56 +37,7 @@ public:
     */
   void renderMap() const;
 
-  /**
- * @brief Sets a node to be highlit
- * @details This sets a node to be highlit, the highlighting is done during rendering
- * @param isoCoordinates which node should be highlit.
- * @param rgbColor The SpriteRGBColor that should be used for highlighting
- */
-  void highlightNode(const Point &isoCoordinates, const SpriteRGBColor &rgbColor);
 
-  /**
- * @brief Sets a node to be unhighlit
- * @details This sets a node to be unhighlit, the highlighting is done during rendering
- * @param isoCoordinates which node should be unhighlit.
- */
-  void unHighlightNode(const Point &isoCoordinates);
-
-  /**
- * @brief Returns the node at given screen coordinates
- *
- * @param screenCoordinates
- * @return Point
- */
-  Point findNodeInMap(const SDL_Point &screenCoordinates, const Layer &layer = Layer::NONE);
-
-  /**
- * @brief Set the Tile ID Of Node object
- * @details Also invokes all necessary texture updates (auto-tiling, slopes, ...)
- * @param tileID the new tileID to set
- * @param coordinate Point where the tileID which should be set
- * @returns setTileID success or not
- */
-  bool setTileID(const std::string &tileID, Point coordinate);
-
-  /**
- * @brief Set the Tile ID Of multiple Node objects
- * @details Also invokes all necessary texture updates (auto-tiling, slopes, ...)
- * @param tileID the new tileID to set
- * @param coordinates a vector of Points where the tileIDs which should be set
- * @returns setTileID success or not
- */
-  bool setTileID(const std::string &tileID, const std::vector<Point> &coordinates);
-
-  /**
- * @brief Demolish a node
- * @details This function gathers all tiles that should be demolished and invokes the nodes demolish function. When a building bigger than 1x1 is selected, all it's coordinates are added to the demolishing points.
- * @param isoCoordinates all coordinates that should be demolished
- * @param updateNeighboringTiles whether the adjacent tiles should be updated. (only relevant for autotiling)
- * @param layer restrict demolish to a single layer
- * @see MapNode#demolishNode
- */
-  void demolishNode(const std::vector<Point> &isoCoordinates, bool updateNeighboringTiles = false, Layer layer = Layer::NONE);
 
   /**
    * @brief Refresh all the map tile textures
@@ -95,51 +46,8 @@ public:
    */
   void refresh();
 
-  /**
-   * @brief Get original corner point of given point within building borders.
-   * @param isoCoordinates Point to get corner point of
-   * @param layer
-   * @returns original corner point
-   */
-  Point getNodeOrigCornerPoint(const Point &isoCoordinates, Layer layer = Layer::NONE);
 
-  /** \brief Save Map to file
-  * @details Serializes the Map class to json and writes the data to a file.
-  * @param fileName The file the map should be written to
-  */
-  void saveMapToFile(const std::string &fileName);
 
-  /** \brief Load Map from file
-  * @details Deserializes the Map class from a json file, creates a new Map and returns it.
-  * @param fileName The file the map should be written to
-  * @returns Map* Pointer to the newly created Map.
-  */
-  static Map *loadMapFromFile(const std::string &fileName);
-
-  /**
- * @brief Debug MapNodeData to Console
- * Used as Tile-Inspector until we implement a GUI variant
- * @param isoCoordinates Tile to inspect
- */
-  void getNodeInformation(const Point &isoCoordinates) const;
-
-  /** \brief check if Tile is occupied
-  * @param isoCoordinates Tile to inspect
-  * @param tileID tileID which should be checked
-  */
-  bool isPlacementOnNodeAllowed(const Point &isoCoordinates, const std::string &tileID) const;
-
-  /** \brief check if Tile can be placed in an area
-  * @param targetCoordinates Tile array to inspect
-  * @param tileID tileID which should be checked
-  */
-  bool isPlacementOnAreaAllowed(const std::vector<Point> targetCoordinates, const std::string &tileID) const;
-
-  /** \brief get Tile ID of specific layer of specific iso coordinates
-  * @param isoCoordinates Tile to inspect
-  * @param layer layer to check.
-  */
-  std::string getTileID(const Point &isoCoordinates, Layer layer);
 
   /** \brief Get pointer to a single mapNode at specific iso coordinates.
   * @param isoCoords The node to retrieve.
@@ -151,24 +59,9 @@ public:
   const std::vector<MapNode> &getMapNodes() { return mapNodes; };
 
 // private:
-  /** \brief Update all mapNodes
-  * @details Updates all mapNode and its adjacent tiles regarding height information, draws slopes for adjacent tiles and
-  * sets tiling for mapNode sprite if applicable
-  */
-  void updateAllNodes();
-
-  /** \brief Get a bitmask that represents same-tile neighbors
-  * @details Checks all neighboring tiles and returns the elevated neighbors in a bitmask:
-  * [ BR BL TR TL  R  L  B  T ]
-  * [ 0  0  0  0   0  0  0  0 ]
-  * @param coordinate Point on the map to calculate bitmask for.
-  * @return Uint that represents a bitmask of the neighbor tiles and their elevation to the center coordinate
-  */
-  std::vector<uint8_t> calculateAutotileBitmask(Point coordinate);
 
   SDL_Color getColorOfPixelInSurface(SDL_Surface *surface, int x, int y) const;
 
-  bool isClickWithinTile(const SDL_Point &screenCoordinates, Point isoCoordinate, const Layer &layer) const;
 
   /** \brief Filter out tiles which should not be set over existing one.
   * @param layer Layer in which tileId should be set.
@@ -176,24 +69,6 @@ public:
   * @return true in case that tileId is allowed to be set, otherwise false (filter it out).
   */
   bool isAllowSetTileId(const Layer layer, const MapNode *const pMapNode);
-
-  /** \brief Get all neighbor nodes from provided map node.
-  * @param isoCoordinates iso coordinates.
-  * @param includeCentralNode if set to true include the central node in the result.
-  * @return All neighbor nodes.
-  */
-  std::vector<NeighborNode> getNeighborNodes(const Point &isoCoordinates, const bool includeCentralNode);
-
-  /** \brief Update the nodes and all affected node with the change.
-  * @param nodes vector of coordinates to be updated.
-  */
-  void updateNodeNeighbors(std::vector<Point> nodes);
-
-  /** \brief Get elevated bit mask of the map node.
-  * @param centerCoordinates Point on the map node to calculate elevated bit mask.
-  * @return Map node elevated bit mask.
-  */
-  unsigned char getElevatedNeighborBitmask(Point centerCoordinates);
 
   /** \brief For implementing frustum culling, find all map nodes which are visible on the screen. Only visible nodes will be rendered.
   */
