+++ src/engine/UIManager.cxx
@@ -528,11 +528,11 @@ void UIManager::setCallbackFunctions()
     }
     else if (uiElement->getUiElementData().actionID == "SaveGame")
     {
-      uiElement->registerCallbackFunction([]() { Engine::instance().saveGame("save.cts"); });
+      // uiElement->registerCallbackFunction([]() { Engine::instance().saveGame("save.cts"); });
     }
     else if (uiElement->getUiElementData().actionID == "LoadGame")
     {
-      uiElement->registerCallbackFunction([]() { Engine::instance().loadGame("save.cts"); });
+      // uiElement->registerCallbackFunction([]() { Engine::instance().loadGame("save.cts"); });
     }
     else if (uiElement->getUiElementData().actionID == "SaveSettings")
     {
